Path relative to source Automatic starting and stopping

+++ python.kak
@@ -1,29 +1,39 @@
 # python bridge for executing things interactively
 
-declare-option str python_bridge_in /tmp/python-bridge-in
-declare-option str python_bridge_out /tmp/python-bridge-out
+declare-option -hidden str python_bridge_in 
+declare-option -hidden str python_bridge_out 
+declare-option -hidden str python_bridge_source %val{source}
+declare-option -hidden bool python_bridge_running false
 
 define-command -docstring 'Create FIFOs and start python -i' \
 python-bridge-start %{
     nop %sh{
         mkfifo $kak_opt_python_bridge_in
         mkfifo $kak_opt_python_bridge_out
-        ( python $kak_config/plugins/kakoune-python-bridge/python-repl.py $kak_opt_python_bridge_in $kak_opt_python_bridge_out 2>/tmp/err) >/dev/null 2>&1 </dev/null &
+        ( python $(dirname $kak_opt_python_bridge_source)/python-repl.py $kak_opt_python_bridge_in $kak_opt_python_bridge_out) >/dev/null 2>&1 </dev/null &
     }
+    set-option global python_bridge_running true
 }
 
 define-command -docstring 'Stop python -i and remove FIFOs' \
 python-bridge-stop %{
     nop %sh{
-        echo "exit()" > $kak_opt_python_bridge_in
-        rm $kak_opt_python_bridge_in
-        rm $kak_opt_python_bridge_out
+        if $kak_opt_python_bridge_running; then
+            cat $kak_opt_python_bridge_out &
+            echo "exit()" > $kak_opt_python_bridge_in
+            rm $kak_opt_python_bridge_in
+            rm $kak_opt_python_bridge_out
+        fi
     }
+    set-option global python_bridge_running false
 }
 
 define-command -docstring 'Evaluate selections using python-bridge' \
 python-bridge-send %{
     evaluate-commands %sh{
+        if ! $kak_opt_python_bridge_running; then
+            echo python-bridge-start
+        fi
         echo "set-register | %{ input=\$(cat); cat $kak_opt_python_bridge_out & echo \"\$input\" > $kak_opt_python_bridge_in & wait}"
     }
     execute-keys -itersel "|<ret>"
@@ -37,4 +47,9 @@ define-command python-bridge -params 1.. -shell-script-candidates %{
 
 hook global KakEnd .* %{
     python-bridge-stop
+}
+
+evaluate-commands %sh{
+    echo "set global python_bridge_in /tmp/kakoune-$kak_session-python-bridge-in"
+    echo "set global python_bridge_out /tmp/kakoune-$kak_session-python-bridge-out"
 }
